pyproject.toml
[project]
name = "youtube-summarizer"
version = "0.1.0"
description = "Download, transcribe, and summarize YouTube videos (Hindi/English → English)"
requires-python = ">=3.12"
dependencies = [
    "mlx-whisper>=0.4.3",
    "ollama>=0.6.1",
    "yt-dlp>=2026.3.17",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.4",
    "itsdangerous>=2.2.0",
    "httpx>=0.27.0",
    "psycopg2-binary>=2.9.0",
    "groq>=1.1.2",
]

[project.scripts]
yt-summarize = "main:main_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = []
